+++ Interface.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#define UP 72
+#define DOWN 80
+#define ENTER 13
+
+
+void StartDraw();
+int MenuDraw();
