+++ main.h
@@ -1,0 +1,20 @@
+#ifndef HEADER_DECLAR  // HEADER_DECLAR라는 매크로가 정의되었는지 확인
+#define HEADER_DECLAR  // 정의안되어 있는 경우 정의(HEADER를 한번만 include하기 위함)
+
+
+#include<iostream>
+#include<cstdio>
+#include<windows.h>
+#include<conio.h>
+#include<time.h>
+
+#define DINO_BOTTOM_Y 12
+#define TREE_BOTTOM_Y 20
+#define TREE_BOTTOM_X 45
+#define UP 72
+#define DOWN 80
+#define ENTER 13
+using namespace std;
+
+
+#endif
