--- project/README.md
@@ -1,1 +1,0 @@
-# 1-Team
