+++ util.h
@@ -1,0 +1,9 @@
+#pragma once
+#include <iostream>
+#include <cstdio>
+#include <windows.h>
+#include <conio.h>
+
+void Init();
+void GotoXY(int x, int y);
+int GetKeyDown();
